// packages/app/src/client/services/Notification/types.ts
/**
 * Copyright 2020 Opstrace, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export type NotificationState = "error" | "warning" | "info" | "success";

export type Notification = {
  id: string;
  title: string;
  information: string;
  state?: NotificationState;
  actions?: NotificationsActions[];
  handleClose?: () => void;
};

export type NotificationsActions = {
  name: string;
  handler: () => void;
};

export type NotificationServiceState = {
  notifications: Notification[];
  visibility: boolean;
};

export type NotificationServiceApi = {
  register: (notification: Notification) => void;
  unregister: (notification: Notification) => void;
  unregisterAll: () => void;
  changeVisibility: () => void;
};